Moved the introductions component to dashboard
