+++ Angular/src/app/app.module.ts
@@ -86,7 +86,7 @@ import { ScansionComponent } from './scansion/scansion.component';
 import { PlaygroundComponent } from './playground/playground.component';
 import { CommentaryComponent } from './commentary/commentary.component';
 import { OverviewComponent } from './overview/overview.component';
-import { IntroductionsComponent } from './introductions/introductions.component';
+import { IntroductionsComponent } from './dashboard/introductions/introductions.component';
 
 // Routes to take. Disallows Path Traversal.
 const appRoutes: Routes = [
